python-bg/requirements.txt
flask
rembg
pillow
watchdog
aiohttp
gradio
numpy
scikit-image